requirements.txt
aiogram>=3.7.0
yt-dlp>=2025.01.12
python-dotenv>=1.0.1
